Update OrderSequenceCounter model to use SMALLINT for primary key to ensure compatibility with Postgres

+++ src/app/models/order/order-sequence-counter.model.ts
@@ -3,7 +3,8 @@ import { Column, DataType, Model, Table } from 'sequelize-typescript';
 /** Single-row counter: `last_assigned` is the last issued display order number (1–100), cycling after 100. */
 @Table({ tableName: 'order_sequence_counter', timestamps: false })
 class OrderSequenceCounter extends Model<OrderSequenceCounter> {
-    @Column({ primaryKey: true, type: DataType.TINYINT }) id: number;
+    /** Postgres has no TINYINT; SMALLINT matches the patch table and range (single row id = 1). */
+    @Column({ primaryKey: true, type: DataType.SMALLINT }) id: number;
 
     @Column({ allowNull: false, type: DataType.SMALLINT })
     last_assigned: number;
